requirements.txt

requests
matplotlib
numpy
python-twitter